Fix UT and FT VNFD yaml because lack of properties
An error occurs in tox due to an incorrect setting of
properties related to tosca.policies.tacker.Scaling.
Therefore, this patch fixes the related VNFDs.
The error is occurring in unittest and FT, so fix both.

At the same time, there was a problem with the version of
heat-translator, so modify the following files as well.
- lower-constraints.txt
- requirements.txt

Co-Authored-By: Yasufumi Ogawa <yasufum.o@gmail.com>
Change-Id: Iae2b25abe53eda483e811171306d3d375a7e1687
Closes-Bug: #1921996