Improvement of Processing Logic for Healing
The following patch will fix the issue wherein if VM name
and AutoScalingGroup name are the same name, unhealthy is
set for both and both are target for healing.

The fix will be in Instantiation part since the root cause
of the issue is that Nova::Server is not included as resource
in vnfcResourceInfo.

Closes-Bug: #1924214
Change-Id: I06849ebb068253e4691b5cfd03caac80bdc16cae