Fix Patch Individual VNF instance
In the condition when the vnf instance hasn't been instantiated,
an error occurs when Patching Individual VNF instance.

1. Fix the exception caused by the NULL value of vnf's
updated_at field.
2. Fix the exception happened when one of its attributes is not
existed in method to_dict in Class VnfInfoModifications.
3. Fix the problem that vnfInstanceName and vnfInstanceDescription
are not displayed in notification when vnfdId is included in the
fields which need modified.
4. In the API reference, modify the method of modifying an
"Individual VNF instance" resource from POST to PATCH.

Closes-Bug: #1915100
Change-Id: Iddb5bb540b817d4fd24e2c7c12f885bd39f8662b