Fix status code for invalid filter conditions
The current tacker returns 500 Internal Server Error to requests with
invalid filter conditions where a VNFID is not UUID. However, as the
error is caused by a bad request, the status code should be 400.

This bug occurs because data types of 'id' and 'vnfd_Id' in TackerObject
are 'string' which don't match the data types in DB, i.e., UUID. In
order to fix this bug, those data types in TackerObject are changed to
'uuid'.

Closes-Bug: 1924666
Closes-Bug: 1926632
Change-Id: I8decbb42d661a44c6c382cb1abbfc37d95084ffb