Add LCM op figures to User Guide
The current document doesn't have an overview of VNF LCM operations,
e.g., Deployment, Scaling and Healing.  For example, the VNF scaling
creates/changes VMs with OpenStack Heat API and users have to confirm
results of an operation with `openstack stack` command, but the
relationship between OpenStack Tacker and OpenStack Heat is not
explicitly described.

For ease of understanding, added diagrams and brief explanation of the
LCM operations listed below.

- ETSI NFV-SOL VNF Deployment as VM with TOSCA
- ETSI NFV-SOL VNF Deployment as VM with LCM operation user data
- ETSI NFV-SOL CNF (Containerized VNF) Deployment
- ETSI NFV-SOL VNF Healing
- ETSI NFV-SOL VNF Scaling
- ETSI NFV-SOL CNF Healing
- ETSI NFV-SOL CNF Scaling

Change-Id: I3ddf9e5ec68c0e84ccc9df4fea9ee274562a3e88