Object definitions of multi version API support
This patch includes only object definitions used by multi version
API support. There is no logic in this patch.

The logic part of multi version API support is provided with a
following patch.

Because the definition of the object had much number of the files,
the patch of logic part is separated to make code review easy.

Implementation notes:
Objects are defined as is according to ETSI NFV-SOL specifications.
This means attributes of objects are not converted to snake case
but camel case as is because the conversion is meaningless
obviously.
Naming convention of class is as follows:
* for a object represents a type with the definition in a section
  of the specication, as is and may be followed by V2 (or V1) to
  distinguish it from another version in the future.
  ex. 'SOL003 5.5.2.2 Type: VnfInstance' => VnfInstanceV2
* for a object representing an inlined struct, 'class name which
  the member belongs to' + '_' + 'member name which changed the
  first letter to the capital letter'.
  ex. the object representing instantiatedVnfInfo in VnfInstance
  => VnfInstanceV2_InstantiatedVnfInfo.
Note that 'nullable=True' means cardinality may be zero in this
object definitions while it means null is OK in the persistent
object originally.

Implements: blueprint multi-version-api
Implements: blueprint support-nfv-solv3-start-and-terminate-vnf
Implements: blueprint support-nfv-solv3-query-vnf-instances
Implements: blueprint support-nfv-solv3-query-operation-occurrences
Implemetes: blueprint support-nfv-solv3-subscriptions
Change-Id: I93711971906132b8dcf794864d22e2c1b326bc3a