Fix retry operation after LCM failure
The LCM retry operation of CNF and MgmtDriver(k8s-cluster)
was verified.

To fix these errors, the following things is done in this patch:
1. When executing retry, if the status of vnf is ERROR, change ERROR to
the status of the corresponding lifecycle.
2. In the scale operation, the judgment condition is added. When vim is
not kubernetes, the `scaleGroupDict` assignment will be used.
3. When an exception occurs in terminate, the task_state in
vnf_instance is changed to None.
4. In the scale operation, add an else branch. If before_error_point is
greater than 4, get resource_changes from vnf_info first. If
resource_changes does not exist, get it by calling the
self._scale_resource_update method.
5. At the beginning of instantiate/heal/terminate, when modifying the
status of vnf, add the judgment condition. When the status is not
ERROR, it can be modified.
6. In the heal/scale operation, add a judgment condition to judge
whether the node exists, and only when the node exists, the delete node
command will be executed.
7. In the instantiate operation, before creating the stack, add to
determine whether the stack already exists.
8. (1) When installing the k8s node with MgmtDriver, add to determine
   whether the node already exists in the cluster. If it exists, skip
   the node and install the next node.
   (2) After installing the k8s cluster, when registering VIM, add to
   determine whether the VIM already exists, and if it exists, use it
   directly.

Closes-Bug: #1942277
Change-Id: If4e695120eb432155499a2e866ae651a18a320ab