Fix duplicate entry error in VNF Package upload
In the current implementation, when we try to upload a VNF package which
has been uploaded and deleted once before, "Duplicate entry" error
occurs and it fails to be onboarded.

This problem happens because VNFD table uses ``vnfd id`` as its primary
key and does not delete a record even if a corresponding VNF package is
deleted.

Although legacy DB is going to be deleted, the patch is not merged at
this time [1]. Thus, this patch modifies the conductor server to simply
delete records from VNFD table.

[1] https://review.opendev.org/c/openstack/tacker/+/763955

Change-Id: I21316cd7b17f85d292a18a44c891f3193ca418ed
Closes-bug: #1920105
Signed-off-by: Hiromu Asahina <hiromu.asahina.az@hco.ntt.co.jp>