support rollback operation task of v2 API
This patch implements rollback operation task defined in ETSI NFV-SOL003
v3.3.1 5.4.15.

This patch supports rollback operation of existing task API, i.e
instantiate. Note that rollback operation of terminate is not supported.

Rollback operation of a task API which is added in the future
(ex. scale) will be implemented in the same patch
of implementing the API.

Functional tests will be provided with another patch.

Implements: blueprint support-nfv-solv3-error-handling
Change-Id: I3b2924287f7b30a85c250602a3f5ecbffd449a5e