Fix VIM registration error
The VIM registration fails in an environment where an admin-endpoint is
close even if a reachable public-endpoint is used in the VIM
configuration.

The cause of this problem is that the endpoint for the keystone API and
the endpoint for the authentication are different inside the keystone
client.  The keystone client implicitly selects admin-endpoint for API
requests from service catalogs unless we specify an endpoint URL or
interface [1], i.e., the auth_url in the VIM configuration isn't used
for API requests.

This patch solves the above problem by specifying an endpoint URL when
creating a keystone client instance.

Note that specifying an interface can be a better solution, as the
keystone can provide different interfaces with the same endpoint URL.
However, to achieve this, we have to add a new field to the VIM
configuration for the interface, which makes a huge modification.
Thus, this patch doesn't take this solution.

Also, this patch includes a tiny refactoring on the OpenStack driver.

[1] https://github.com/openstack/python-keystoneclient/blob/d5cb761763988217cdd00a7981a62e2083fd381f/keystoneclient/httpclient.py#L251

Change-Id: Ic743fadbd1ddd59ca6755949e03b77c8715be459
Closes-bug: #1920088
Signed-off-by: Hiromu Asahina <hiromu.asahina.az@hco.ntt.co.jp>