CI: Create admin endpoint for keystone
A recent devstack change [1] has switched off the creation of an admin
endpoint for keystone, which we need. Create it again until we update
Tacker to stop using it.

And devstack patch [2] has stopped creating userrc_early and If we do
not use OS_CLOUD, we must override the --os-cloud option of the
openstack command.
So, this patch fixes to set default OS_CLOUD setting overridden.

And for UUID formatting checks in db/types.py, fix to use first
element for List, since value may be passed as a List.

[1] https://review.opendev.org/c/openstack/devstack/+/735472
[2] https://review.opendev.org/c/openstack/devstack/+/780417

Change-Id: I39be7c48aca4be9a4e0bcec6dc6fa45b11c623b3