Fix some bugs related Kubernetes MgmtDriver
After the X version was released, regression tests were performed
on all functions of MgmtDriver. This patch is used to fix
these bugs.

1. Add the process of getting 'FloatingIP' to kuberenetes_mgmt.py
and private_registry_mgmt.py files, so that users can access the
Registry and Storage VM through the floatingIP when using the CIR
and Storage functions.
2. When the registry is accessed through https, add a `break`
process to the while loop connection in the
`_connect_to_private_registries` method in kuberenetes_mgmt.py
to break out of the loop.
3. Modify the type field in podaffinity_hot_top.yaml to match
the name in the nested directory.
4. When using the pod-affinity function, if multiple
'OS::Nova::ServerGroup' resources are created at the same time,
the id of the resource cannot be matched to the workerNode.
Modify the original process of obtaining the id directly,
instead, first obtain the `physical_resource_id` of all
'OS::Nova::ServerGroup' resources, and then match the id in the
workerNode information to obtain the target id.
5. Modify the method of obtaining the kubespray VM's password
in kubespray_mgmt.py.

Closes-Bug: #1949926
Change-Id: I2e966df35f874e14c39792a6882ee913ac424663