Fix filter data creation
The tacker-server returns 500 error if a request body of the
subscription creation [1] that doesn't have
`vnfInstanceSubscriptionFilter` parameter, although it's an optional
parameter.

The happens simply because the `vnfInstanceSubscriptionFilter` is not
handled as an optional parameter in the current codes.

This patch fixes such codes to correctly handle the request without a
`vnfInstanceSubscriptionFilter` parameter.

[1]:
https://docs.openstack.org/api-ref/nfv-orchestration/v1/vnflcm.html?expanded=create-a-new-subscription-detail

Change-Id: Ibf9310c4ddffcc41e23eea8c5dfd212fd3a9a145
Closes-bug: #1947303