Test to create VNF instance with 'VNF' name
Sometimes API layer can add restrictions to not allow
resource creation with the resource code name like vnf
creation with 'VNF' itself. There is no such restriction
in tacker and we can test it explicitly.

This commit modifies an existing test to create vnf with
the name 'VNF' which will verify the things. We do not need
to add a separate test for that as it can be verified in one
of the existing tests itself.

Change-Id: I41ea57ef9a5c35c124d4fa43f8c3b9498bd80c2c