CI: Extend timeout for CNF entire heal in Helm FT
An error occurred at the FT of CNF LCM with Helm chart on the
sol-kubernetes job due to the time it took to remove the Kubernetes
resources for the entire heal operation.
There is not enough time to wait for the status to COMPLETED, and the
heal operation itself succeeded.
This patch extends the timeout period to avoid errors.

Change-Id: I1a39707e64c481677c9bad9a6c3073dc2b9ed6ce