Increase stack_retries of k8s FT in Zuul
Due to the performance of the zuul environment, it may take a long time
to remove the kubernetes resource and an error may occur.
This patch avoids errors by increasing the
`kubernetes_vim.stack_retries` value in Tacker config and changes the
timeout to value calculated from config.
This patch also contains a fix for a test where sol003 heal ran twice
in Helm test.

Change-Id: I0fc8d6748cf92f6384b6cdc49322f0f85f16afcd