Change test utils to use absolute path
In the current implementation, a relative path from the project root is
used to get sample files (e.g., VNF packages) for tests. Due to this,
it's not possible to run a specific test in different place from the
project root.

This patch solves this problem by obtaining absolute paths from a test
file.

Change-Id: I541ea67965fc24121f0dca6aabd9d4b9210d9175