Refactor constants in conductor
In the current implementation, some groups of constant values are
defined in conductor_server as a tuple that is a dependency on
`_change_vnf_status`. There are two problems: (i) the place where the
constants declared is inappropriate; (ii) it depends on a specific
function.

This patch solves these problem by moving constants declaration to
constants.py and modify the `_change_vnf_status`.

Change-Id: I7255e18f95ac3fd8199d25181b7b3c4f9de01b51