Suppress Warnings for SQL Alchemy
The ``cache_ok`` attribute for SQL Alchemy TypeDecorator will emit a
warning when its value is set to the default value ``None``. When this
value is set to True, SQL Alchemy will cache SQL query and reuse it
later when the same query is called again, and vice versa.

This patch simply sets ``cache_ok`` to TypeDecorator classes according
to whether it can be used as a cache key or not [1].

[1] https://docs.sqlalchemy.org/en/14/core/type_api.html#sqlalchemy.types.ExternalType.cache_ok

Closes-Bug: #1921981
Change-Id: I503c93d6e8869b30c49909c95fb2669862d28a3f