Introduce vagrant for devstack
This update is for providing a dedicated devstack installer for Tacker.
It is configurable for supporting several usecases, such as multiple
nodes configuration. It's also supports both of default distros in
OpenStack, Ubuntu 20.04 (focal) and CentOS Stream 8 currently.
You can find usages in the contributor guide `Devstack Installation
with Vagrant`.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>
Implements: bp intro-tacker-installer
Change-Id: I2f9fcf0811c41d9278df79b5a7e60d608d23cd76