Clean up at conductor start up
Previously when the conductor fails down during operation of
some vnf instances, operationState of VnfLcmOpOcc is left as
STARTING, PROCCESING or ROLLING_BACK.
As a result, no more operation of the vnf instance can't be processed.

This patch fixes this situation. When the conductor (re) starts up,
if VnfLcmOpOcc with its operationState is STARTING, PROCCESING or
ROLLING_BACK exists, it is changed to ROLLED_BACK or FAILED_TEMP.

Implements: blueprint support-nfv-solv3-error-handling
Change-Id: Ib50869b8b32bfb9caaa4b0b95236b33d04dfb2fe