Add a Sample VNF Package for Ansible Driver
This patch contains a sample VNF Package that has ansible driver
artifacts. It will be a useful reference for users who choose to use
ansible to configure their target VNFs.

User document of ansible driver usage will be added.
It describes the conditions for running the script, how to create the
package, and an example of how it works.

Implements: blueprint add-ansible-mgmt-driver-sample
Spec: https://review.opendev.org/c/openstack/tacker-specs/+/814689
Change-Id: Ibbb9115a938de19e48f06c59af6bd7e7b222706d