Fix usageState of VNF package update
A VNF package can be associated with multiple VNF instances.
Previously when a VNF instance is deleted, usageState of the
associated VNF package is changed to 'NOT_IN_USE' unconditionally.
It is not correct if there is another VNF instance associated
same VNF package.
This patch fixes to change usageState to 'NOT_IN_USE' only if
there is no VNF instance associated the VNF Package.

Closes-Bug: #1963858
Change-Id: Ib15ffd89d057079a50e6ebc62378e0d83086b107