Fix the definition of associated_vnfc_cp_id
This patch will fix the definition of the field of
associated_vnfc_cp_id. In the current implementation,
the field of associated_vnfc_cp_id is not allowed to be null.
But according to ETSI GS NFV-SOL 003 V3.6.1, the cardinality of
associatedVnfcCpId is defined as "0..1",
which means the associated_vnfc_cp_id is nullable.
Therefore, this patch will fix it to comply with the standard of ETSI.

Closes-Bug: #1948925
Change-Id: I612835a7f8dd61c034b0f877d3356c901fc6027f