Use full URL to call external NFVO
Common http utilities used by v2 API use keystoneauth1 module.
If a server is a service in OpenStack, clients can omit endpoint.
But if a server is a service outside of OpenStack, clients can't
omit endpoint.
This patch fixes to add endpoint to URL in case that a server is
an external NFVO.

Change-Id: I3cf8b608d2996a47f798184ae4006282f672dd02