Fix failure of testing privsep
Unittests for privsep and rootwrap fail if it runs on non-linux env
such as MacOS, or other than devstack environment. It's because
`oslo.privsep` and its tool are not installed properly in tox env.
So, skip the tests on other than devstack env to avoid such a failure.

This patch also modifies `pypath` in initialization of Privcontext
because its fstring cannot be expanded correctly in some environment
and tests are failed.

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>
Change-Id: I1e160500c30c379645bc19a67c429e9b2f12dcda