Enable multiple VNFs creation from one package
This patch removes the restriction that only one VNF can be created from
a single VNF package when deploying CNFs.
The process of associating the `VDU.properties.name` in VNFD with the
actual resource name has changed by adding the `vdu_mapping` parameter
to InstantiateVnfRequest.additionalParams.

The patch also updates the documentation to remove the restriction and
add how to instantiate with `vdu_mapping` parameter.

Implements: blueprint remove-cnf-restriction
Change-Id: I32c041b8fd996b9b338fe97eb1604567fd6b6aaf