Implemented check on range end to fetch artifact
Getting contents of the file instead of 416 error when
range end is greater than or equal to size of the file.

Added a check for range end while fetching artifact
information. If range end is greater than or equal to
size of the file then 416 error will be returned in
response.

Closes-Bug: #1959885
Change-Id: I8dbf7f8407f199350476fc7f093a3cccc78e5030