Fix k8s init error when file contains special key
When executing CNF lifcycle operations, if the manifest file
contains the following parameters:
  `exec`, `not`, `except`, `continue`, `from`
the initialization of Kubernetes resources will fail.

To fix this issue, this patch makes changes to the `_init_k8s_obj`
method in the `translate_outputs.py` file. When the manifest file
contains the above parameters, it is first converted into
parameters that can be recognized by the k8s client module and
then initialized.

Closes-Bug: #1978071
Change-Id: I392e1b07152b260de775754dbe4606995f658d8d