Update config in tacker installer
For a limitation of IP address range of hostonly network which was
added in the recent update in VirtualBox [1], allowed only
192.168.56.0/21, the installer is terminated without you give
appropriate IP addresses to your VMs.

Forwarding port `10080` for HTTP is also invalid in security for some
browsers. `20080` is valid for the restriction.

This update is to update default values of IP address and port to be
allowed.

[1] https://www.virtualbox.org/manual/ch06.html#network_hostonly

Closes-Bug: #1980312
Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>
Change-Id: I7fbce874ce2445ea72e627e0ad6309062fac9292