Improve Unit Test for V2 API
In the existing v2 unit test, there are many public methods that
are not tested. This led to bugs in some code and also increased
rework in FT.

This patch adds tests to some public methods and improves some
existing tests. Also fixed several minor issues checked out in UT.

Change-Id: I15b45d99646babbb57606dbb46aa156aa9dfb51c