Refactor change_vnfpkg v2 API
This patch refactors the implementation of change_vnfpkg v2 API.

Because the original implementation of change_vnfpkg does not obey
the rules that other operations obey, it makes hard to enhance
the function that related to whole operations such as the HOT format
change. This patch corrects it.

This patch also changes some specifications of the API and fixes
some bugs (described in bug #1981533).

Specification changes:
* image to volume translation (or vice versa) is not supported.
  Note that it did not work property by a bug.
* parameters to the coordinate script are changed.
* property name for images in the definition of AutoScalingGroup is
  changed.
* unnecessary parameters of additionalParams,
  'lcm-operation-coordinate-old-vnf-class' and
  'lcm-operation-coordinate-new-vnf-class' are removed.
* 'lcm-operation-coordinate-old-vnf' and
  'lcm-operation-coordinate-new-vnf' in additionalParams are option
   now.

Detail of bug fiexes are omitted.

Note that the object of this patch is common part and openstack VIM part.
k8s VIM part will be fixed by another patch.

Closes-Bug: #1981533
Change-Id: I9f6d103d66e6904704ca467d68c4ebfbcbd8a1f8