Fix IPv4 check failure in UT
Since the version of jsonschema has been updated from 3.2.0 to 4.7.2,
when tox executes UT, it reports an error when checking IPv4, and now
the invalid IP address is fixed.

Due to FT failure caused by ceilometer, the installation of Ceilometer
was commented out in `.zuul.yaml`.
Note that this is only a temporary change and will need to be changed
back when the ceilometer issue is resolved.

Closes-Bug: #1983067
Change-Id: Ice9be4d71ee1689b57add263be5fadd77a7f5ccd