Enable retries when sending notifications
This patch enables retries when a connection error occurred
at the time of sending a notification.

This patch introduces a new config parameter
v2_vnfm.notify_connect_retries that specifies the maximum number
of retries. Note that the period between retries is exponential
starting 0.5 seconds up to a maximum of 60 seconds.

Change-Id: Ie75a7d1ec7a8dd8970b1cf58cff86f699bcc3311