Fix instantiate failed after terminate operation
After the terminate operation is performed, the `instantiatedVnfInfo`
field in vnf_instance will retain some attributes, causing the method
to obtain the `stack_name` to fail when the instantiate operation
is performed again.

This patch modifies the logical judgment in the existing
`get_stack_name` method and fixes the above issue.

Closes-Bug: #1991888
Change-Id: I6863d4e400a8770b4527401c7c9e49851220cba5