Enable vnfcInfo id to be fixed
This patch enables vnfcInfo id to be fixed as a part of
individual vnfc management. A user can specify vnfcInfo ids
by using metadata of VnfInstance in advance.

The format of specifying vnfcInfo ids as follows:
```
    "VDU_VNFc_mapping": {
        <VDU name>: [<list of vnfcInfo ids>],
        ...
    }
```

If this is not in the metadata of VnfInstance, vnfcInfo ids are
set as before.

Implements: blueprint individual-vnfc-management
Change-Id: I93aed376b1d355f9dcfb07a7711556620efeae77