Run Tacker gate job in Ubuntu Focal
As per 2023.1 testing runtime[1], we need to test on Ubuntu
Jammy and at least run a single job on Ubuntu Focal(for smooth
upgrade).

This patch adds a new sol-v2 tacker job to run on Focal which
can be removed in the future cycle when testing runtime drops the
requirement of Focal testing.

[1] https://governance.openstack.org/tc/reference/runtimes/2023.1.html

Change-Id: I96a611dabc947a876c978a55208df48b59ec6421