Remove VIM monitoring depending on Mistral
This patch is a part of Legacy APIs deprecation.

Currently, the Tacker uses Mistral for the VIM monitoring but for
nothing else. Maintaining dependency on Mistral only for this feature
is ineffective in terms of maintainability. Moreover, the main use case
of Tacker is now VNFM that is not responsible for VIM monitoring.

This patch removes the dependency on Mistral for VIM monitoring.

Note: although there's no actual use, Tacker also depends on Mistral in
NS LCM, which will be removed by a different patch.

Implements: blueprint deprecate-legacy-apis

Change-Id: Ia11496da7942457239df59b52af817c76e736013