Missing Location in http header
This patch fixes the problem regarding the "Location" HTTP header
which is missing in the response when changing external connectivity.
According to the specification file, When operating on VNF instance
resource.

The VNFM returns a "202 Accepted" response with an empty payload body
and a "Location" HTTP header that points to the new "Individual VNF LCM
operation occurrence" resource, i.e. it includes the URI of that
resource which is "…/vnf_lcm_op_occs/{vnfLcmOpOccId}" but now "Location"
HTTP header is missing in the response when changing external
connectivity.

Closes-Bug: #1945387
Change-Id: I77c161b08c9c468fe6d4d243c1d098a0855bb2c8