Support CNF auto scale via PM Threshold interface
Add the Performance Management Threshold API and CLI to support
AutoScale. The Performance Management Threshold API is based on
ETSI NFV-SOL 002 v3.3.1 and ETSI NFV-SOL 003 v3.3.1, on which v2
VNF LCM API of Tacker is based. Add the Prometheus Plugin that
has an interface between Tacker and Prometheus as a sample of
External Monitoring Tool.

Implements: blueprint support-auto-lcm
Change-Id: If41e89c5601592aea18a81df54d0db44a8588e1c