Fix missing teardown in PrometheusPluginTest
In the existing codes, a singleton object of PrometheusPluginDriver
remains after tests are finished. This can incur other tests failure as
most tests assume that the instance is absent, and therefore, a stub
for the plugin is used instead [1].

This patch fixes invalid test cases to ensure deleting a singleton
object of PrometheusPlugin in the tearDown phase.

[1]
https://github.com/openstack/tacker/blob/master/tacker/sol_refactored/common/prometheus_plugin.py/#L62

Closes-Bug: #2007289
Change-Id: Ie92e4b82b913c6386057a55d7e63e5586d5983b6