Support importing def from local instead of url
Sometimes importing a nested definition from remote server is failed
for the server is busy. Although we usually can download the definition
before and change the path as local then, cannot for a case if the
contents of the template is fixed and it should not be changed.

This update is to avoid such a case. The url can be replaced with any
other path by using arg `local_defs` like as below. In this case, the
definition is imported from local path instead of url. It's also
includes some util functions to make it easier developers to find local
files.

    local_defs = {"https://example.org/orig": "/path/to/local/def"}
    tosca = ToscaTemplate(template, local_defs=local_defs}

Closes-Bug: #2008048
Depends-On: https://review.opendev.org/c/openstack/tosca-parser/+/873217
Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>
Change-Id: Ied30614f22282a85badb568854f20f2b76f64d70