Add user docs for Coordination API sample usage
This patch adds a sample usage for the client
of ETSI NFV-SOL VNF Coordination API
in the coordinateVNF script to user guide.

User docs provides how to use the Coordination API in the
coordinateVNF script and the following operations.
* Request a coordination action
  (POST /lcmcoord/v1/coordinations)
* Read the result of a coordination action
  (GET /lcmcoord/v1/coordinations/{coordinationId})

Note that the sample script implements the client
function in the ChangeCurrentVNFPackage API only,
not the VNF LCM Coordination API itself.

Implements: blueprint add-sample-coordinate-script
Change-Id: I789e774659e9a54e47c7b24161fddfffc4549fcf